LayoutTests/fast/dom/CSSStyleDeclaration/css-properties-case-sensitive.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
'This test checks that access to CSS properties via JavaScript properties on DOM elements is case sensitive.'
);

var element = document.createElement('a');
element.style.zIndex = 1;

shouldBe("element.style.zIndex", "'1'");
shouldBeUndefined("element.style.ZIndex");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
