LayoutTests/fast/dom/CSSStyleDeclaration/css-properties-case-sensitive-expected.txt

This test checks that access to CSS properties via JavaScript properties on DOM elements is case sensitive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.style.zIndex is '1'
PASS element.style.ZIndex is undefined.
PASS successfullyParsed is true

TEST COMPLETE
